Update currency.js
Changes to key press handling now allows changes to be made where the cursor is located or to replace selected elements. Also if the cursor is not after the decimal then numbers can be added but if the cursor is past the decimal and a character is attempted it will only be allowed if there are less than two numbers after the decimal.